requirements.txt
chromadb==0.3.20
guidance==0.0.64
langchain==0.0.234
numpy==1.23.5
openai==0.28.0
pydantic==1.10.7
python-dotenv==0.19.1
streamlit==1.23.1